fix merge conflicts, lint
